sqs_consumer/requirements.txt
boto3
python-dotenv
gspread
google-auth 